requirements.txt
Django==4.1.7
psycopg2==2.9.5
Pillow==9.4.0
djangorestframework==3.14.0